moved swagger into its own router
